auth/requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.6.0
async-fastapi-jwt-auth==0.6.6
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.5
fastapi-jwt==0.3.0
fastapi_pagination==0.12.31
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
PyYAML==6.0.2
redis==5.0.4
rich==13.8.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.37.2
typer==0.12.5
typing_extensions==4.12.2
user-agents==2.2.0
ujson==5.10.0
uvicorn==0.30.6
# uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
Alembic==1.13.3
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
itsdangerous==2.2.0